[Fuzzing] Updated the LSR fuzzers to fuzz the public interfaces and moved the fuzzing mod.
Closes: #5799
