+++ consensus/safety-rules/src/fuzzing_utils.rs
@@ -252,13 +252,62 @@ pub fn arb_safety_rules_input() -> impl Strategy<Value = SafetyRulesInput> {
     ]
 }
 
+#[cfg(any(test, feature = "fuzzing"))]
+pub mod fuzzing {
+    use crate::{error::Error, serializer::SafetyRulesInput, test_utils, TSafetyRules};
+    use consensus_types::{
+        block::Block, block_data::BlockData, timeout::Timeout, vote::Vote,
+        vote_proposal::MaybeSignedVoteProposal,
+    };
+    use libra_crypto::ed25519::Ed25519Signature;
+    use libra_types::epoch_change::EpochChangeProof;
+
+    pub fn fuzz_initialize(proof: EpochChangeProof) -> Result<(), Error> {
+        let mut safety_rules = test_utils::test_safety_rules();
+        safety_rules.initialize(&proof)
+    }
+
+    pub fn fuzz_construct_and_sign_vote(
+        maybe_signed_vote_proposal: MaybeSignedVoteProposal,
+    ) -> Result<Vote, Error> {
+        let mut safety_rules = test_utils::test_safety_rules();
+        safety_rules.construct_and_sign_vote(&maybe_signed_vote_proposal)
+    }
+
+    pub fn fuzz_handle_message(safety_rules_input: SafetyRulesInput) -> Result<Vec<u8>, Error> {
+        // Create a safety rules serializer test instance for fuzzing
+        let mut serializer_service = test_utils::test_serializer();
+
+        // LCS encode the safety_rules_input and fuzz the handle_message() method
+        if let Ok(safety_rules_input) = lcs::to_bytes(&safety_rules_input) {
+            serializer_service.handle_message(safety_rules_input)
+        } else {
+            Err(Error::SerializationError(
+                "Unable to serialize safety rules input for fuzzer!".into(),
+            ))
+        }
+    }
+
+    pub fn fuzz_sign_proposal(block_data: BlockData) -> Result<Block, Error> {
+        let mut safety_rules = test_utils::test_safety_rules();
+        safety_rules.sign_proposal(block_data)
+    }
+
+    pub fn fuzz_sign_timeout(timeout: Timeout) -> Result<Ed25519Signature, Error> {
+        let mut safety_rules = test_utils::test_safety_rules();
+        safety_rules.sign_timeout(&timeout)
+    }
+}
+
 // Note: these tests ensure that the various fuzzers are maintained (i.e., not broken
 // at some time in the future and only discovered when a fuzz test fails).
 #[cfg(test)]
 mod tests {
     use crate::{
-        fuzz_construct_and_sign_vote, fuzz_handle_message, fuzz_initialize, fuzz_sign_proposal,
-        fuzz_sign_timeout,
+        fuzzing::{
+            fuzz_construct_and_sign_vote, fuzz_handle_message, fuzz_initialize, fuzz_sign_proposal,
+            fuzz_sign_timeout,
+        },
         fuzzing_utils::{
             arb_block_data, arb_epoch_change_proof, arb_maybe_signed_vote_proposal,
             arb_safety_rules_input, arb_timeout,
