+++ consensus/safety-rules/src/safety_rules.rs
@@ -458,38 +458,3 @@ where
             err
         })
 }
-
-#[cfg(any(test, feature = "fuzzing"))]
-/// This module allows the fuzzer to fuzz the guarded_* request handlers used by safety rules to
-/// handle message requests.
-pub mod fuzzing {
-    use crate::{error::Error, test_utils};
-    use consensus_types::{
-        block::Block, block_data::BlockData, timeout::Timeout, vote::Vote,
-        vote_proposal::MaybeSignedVoteProposal,
-    };
-    use libra_crypto::ed25519::Ed25519Signature;
-    use libra_types::epoch_change::EpochChangeProof;
-
-    pub fn fuzz_initialize(proof: EpochChangeProof) -> Result<(), Error> {
-        let mut safety_rules = test_utils::test_safety_rules();
-        safety_rules.guarded_initialize(&proof)
-    }
-
-    pub fn fuzz_construct_and_sign_vote(
-        maybe_signed_vote_proposal: MaybeSignedVoteProposal,
-    ) -> Result<Vote, Error> {
-        let mut safety_rules = test_utils::test_safety_rules();
-        safety_rules.guarded_construct_and_sign_vote(&maybe_signed_vote_proposal)
-    }
-
-    pub fn fuzz_sign_proposal(block_data: BlockData) -> Result<Block, Error> {
-        let mut safety_rules = test_utils::test_safety_rules();
-        safety_rules.guarded_sign_proposal(block_data)
-    }
-
-    pub fn fuzz_sign_timeout(timeout: Timeout) -> Result<Ed25519Signature, Error> {
-        let mut safety_rules = test_utils::test_safety_rules();
-        safety_rules.guarded_sign_timeout(&timeout)
-    }
-}
